:sparkles: Validação de usuário logado para a deleção

+++ src/main/java/com/futurodev/ativa365/controller/PersonController.java
@@ -3,6 +3,7 @@ package com.futurodev.ativa365.controller;
 import com.futurodev.ativa365.exceptions.PersonCpfAlreadyExistsException;
 import com.futurodev.ativa365.exceptions.PersonEmailAlreadyExistsException;
 import com.futurodev.ativa365.exceptions.PersonNotFoundException;
+import com.futurodev.ativa365.exceptions.PersonToBeDeletedIsNotTheCurrentUser;
 import com.futurodev.ativa365.model.transport.CreatePersonForm;
 import com.futurodev.ativa365.model.transport.PersonDTO;
 import com.futurodev.ativa365.service.PersonService;
@@ -11,10 +12,13 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.util.UriComponentsBuilder;
 
 import java.net.URI;
+import java.util.Objects;
 
 @RestController
 @RequestMapping("/usuario")
@@ -41,9 +45,16 @@ public class PersonController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<PersonDTO> deletePerson(@PathVariable("id") Long id)
-            throws PersonNotFoundException {
-        this.personService.deletePerson(id);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<PersonDTO> deletePerson(@PathVariable("id") Long id,
+                                                  @AuthenticationPrincipal UserDetails userInSession)
+            throws PersonNotFoundException, PersonToBeDeletedIsNotTheCurrentUser {
+        PersonDTO personInSession = this.personService.getUserInSession(userInSession.getUsername());
+        if(Objects.equals(personInSession.id(), id)){
+            this.personService.deletePerson(id);
+            return ResponseEntity.noContent().build();
+        } else{
+            throw new PersonToBeDeletedIsNotTheCurrentUser(id);
+        }
+
     }
 }
